#feat -- random hero updates
