+++ styles/raw-fonts.scss
@@ -690,3 +690,236 @@
     url('~/assets/fonts/Twitch/twitch.ttf') format('truetype'),
     url('~/assets/fonts/Twitch/twitch.otf') format('truetype');
 }
+
+// matter light
+@font-face {
+  font-family: 'Matter';
+  font-style: normal;
+  font-weight: 300;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-Light.otf') format('truetype');
+}
+
+// matter light italic
+@font-face {
+  font-family: 'Matter';
+  font-style: italic;
+  font-weight: 300;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-LightItalic.otf') format('truetype');
+}
+
+// matter regular
+@font-face {
+  font-family: 'Matter';
+  font-style: normal;
+  font-weight: 400;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-Regular.otf') format('truetype');
+}
+
+// matter regular italic
+@font-face {
+  font-family: 'Matter';
+  font-style: italic;
+  font-weight: 400;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-RegularItalic.otf') format('truetype');
+}
+
+// matter medium
+@font-face {
+  font-family: 'Matter';
+  font-style: normal;
+  font-weight: 500;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-Medium.otf') format('truetype');
+}
+
+// matter medium italic
+@font-face {
+  font-family: 'Matter';
+  font-style: italic;
+  font-weight: 500;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-MediumItalic.otf') format('truetype');
+}
+
+// matter semibold
+@font-face {
+  font-family: 'Matter';
+  font-style: normal;
+  font-weight: 600;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-SemiBold.otf') format('truetype');
+}
+
+// matter semibold italic
+@font-face {
+  font-family: 'Matter';
+  font-style: italic;
+  font-weight: 600;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-SemiBoldItalic.otf') format('truetype');
+}
+
+// matter bold
+@font-face {
+  font-family: 'Matter';
+  font-style: normal;
+  font-weight: 700;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-Bold.otf') format('truetype');
+}
+
+// matter bold italic
+@font-face {
+  font-family: 'Matter';
+  font-style: italic;
+  font-weight: 700;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-BoldItalic.otf') format('truetype');
+}
+
+// matter heavy
+@font-face {
+  font-family: 'Matter';
+  font-style: normal;
+  font-weight: 800;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-Heavy.otf') format('truetype');
+}
+
+// matter heavy italic
+@font-face {
+  font-family: 'Matter';
+  font-style: italic;
+  font-weight: 800;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/Matter/Matter-HeavyItalic.otf') format('truetype');
+}
+
+
+// ANTIQUE OLIVE
+
+// antique olive light
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: normal;
+  font-weight: 300;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-Light.ttf') format('truetype');
+}
+
+// antique olive light italic
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: italic;
+  font-weight: 300;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-LightItalic.ttf') format('truetype');
+}
+
+// antique olive regular
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: normal;
+  font-weight: 400;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-Regular.ttf') format('truetype');
+}
+
+// antique olive regular italic
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: italic;
+  font-weight: 400;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-RegularItalic.ttf') format('truetype');
+}
+
+// antique olive medium
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: normal;
+  font-weight: 500;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-Medium.ttf') format('truetype');
+}
+
+// antique olive medium italic
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: italic;
+  font-weight: 500;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-MediumItalic.ttf') format('truetype');
+}
+
+// antique olive Bold
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: normal;
+  font-weight: 700;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-Bold.ttf') format('truetype');
+}
+
+// antique olive Bold condensed
+// @font-face {
+//   font-family: 'Antique Olive';
+//   font-style: normal;
+//   font-weight: 700;
+//   font-display: swap;
+//   src:
+//     url('~/assets/fonts/antique-olive/AntiqueOlive-BoldCondensed.ttf') format('truetype');
+// }
+
+// antique olive heavy
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: normal;
+  font-weight: 900;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-Heavy.ttf') format('truetype');
+}
+
+// antique olive heavy italic
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: italic;
+  font-weight: 900;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-HeavyItalic.ttf') format('truetype');
+}
+
+// antique olive black
+@font-face {
+  font-family: 'Antique Olive';
+  font-style: normal;
+  font-weight: 800;
+  font-display: swap;
+  src:
+    url('~/assets/fonts/antique-olive/AntiqueOlive-Black.ttf') format('truetype');
+}
