+++ styles/typography.scss
@@ -12,7 +12,7 @@
 }
 
 $fonts-map: (
-  "sans-serif": "Calibre",
+  "sans-serif": "Matter",
   // "sans-serif": "Brown",
   // "sans-serif": "DM Sans",
   // "sans-serif": "Brandon Grotesk",
@@ -32,7 +32,13 @@ $fonts-map: (
 
   "rubik-puddles": "Rubik Puddles",
 
-  "twitch": "Twitch"
+  "twitch": "Twitch",
+
+  "matter": "Matter",
+
+  "big-heading": "Antique Olive",
+
+  "antique-olive": "Antique Olive"
 );
 
 $font-sizes: (
